requirements.txt
Pillow>=10.0.0,<11.0.0
flask>=2.0.0
flask-cors>=4.0.0
